Switch from ENTRYPOINT to CMD Dockerfile and clean up
They are similar but CMD is more appropriate because it allows
overriding on the command line, e.g. `docker run ... kill-sticky sh`
can now be used to start an interactive shell for debugging.

Also remove `set -x` from Dockerfile (unnecessary because we use
`&&`), and switch from `$(pwd)` to `$PWD` in the build commands
because `$PWD` is cleaner (especially in the Makefile).